--- tests/test_vis.py
@@ -1,268 +1,0 @@
-"""Comprehensive tests for NRDK visualization module."""
-
-import numpy as np
-import torch
-
-from nrdk.vis import (
-    bev_from_polar2,
-    bev_height_from_polar_occupancy,
-    depth_from_polar_occupancy,
-    swap_angular_conventions,
-    tile_images,
-)
-
-
-def _small_rgb_images():
-    """Generate small RGB test images (2,3,5,3)."""
-    torch.manual_seed(42)
-    return torch.randn(2, 3, 5, 3)
-
-
-def _small_grayscale_images():
-    """Generate small grayscale test images (2,3,5)."""
-    torch.manual_seed(42)
-    return torch.randn(2, 3, 5)
-
-
-def _small_polar_occupancy():
-    """Generate small polar occupancy grids (2,3,6,8)."""
-    torch.manual_seed(42)
-    return torch.randint(0, 2, (2, 3, 6, 8), dtype=torch.bool)
-
-
-def _small_polar_data():
-    """Generate small polar data (2,6,8,1)."""
-    torch.manual_seed(42)
-    return torch.randn(2, 6, 8, 1)
-
-
-# Utils Tests
-
-def test_tile_images_grayscale_basic():
-    """Test tile_images with grayscale images."""
-    images = _small_grayscale_images()  # (2,3,5)
-    result = tile_images(images, cols=2, normalize=True)
-
-    # Should create 1 row × 2 cols grid: (3, 10, 3)
-    assert result.shape == (3, 10, 3)
-    assert result.dtype == np.uint8
-
-
-def test_tile_images_rgb_basic():
-    """Test tile_images with RGB images."""
-    images = _small_rgb_images()  # (2,3,5,3)
-    result = tile_images(images, cols=2, normalize=True)
-
-    # Should create 1 row × 2 cols grid: (3, 10, 3)
-    assert result.shape == (3, 10, 3)
-    assert result.dtype == np.uint8
-
-
-def test_tile_images_multiple_predictions():
-    """Test tile_images with y_true and y_hat."""
-    y_true = _small_grayscale_images()  # (2,3,5)
-    y_hat = _small_grayscale_images() * 0.5  # Different values
-
-    result = tile_images(y_true, y_hat, cols=2, normalize=True)
-
-    # Should stack vertically: y_true on top, y_hat below
-    # Result: (6, 10, 3) - doubled height
-    assert result.shape == (6, 10, 3)
-
-
-def test_tile_images_padding():
-    """Test tile_images with non-divisible batch size."""
-    # Create 3 images but request 2 columns
-    images = torch.randn(3, 3, 5)
-    result = tile_images(images, cols=2, normalize=True)
-
-    # Should pad to 4 images (2 rows × 2 cols): (6, 10, 3)
-    assert result.shape == (6, 10, 3)
-
-
-def test_tile_images_custom_colormap():
-    """Test tile_images with custom colormap array."""
-    images = _small_grayscale_images()  # (2,3,5)
-    result = tile_images(images, cols=1, cmap='plasma', normalize=True)
-
-    assert result.shape == (6, 5, 3)
-    assert result.dtype == np.uint8
-
-
-def test_tile_images_no_normalization():
-    """Test tile_images without normalization."""
-    # Pre-normalized data in [0, 1]
-    images = torch.rand(2, 3, 5)
-    result = tile_images(images, cols=2, normalize=False)
-
-    assert result.shape == (3, 10, 3)
-
-
-def test_swap_angular_conventions_basic():
-    """Test basic swap_angular_conventions functionality."""
-    data = torch.arange(24).reshape(2, 3, 4)
-    result = swap_angular_conventions(data)
-    assert result.shape == (2, 4, 3)
-
-    # Verify it's not just a transpose (would miss the flip)
-    simple_transpose = data.transpose(-1, -2)
-    assert not torch.allclose(result, simple_transpose)
-
-
-def test_swap_angular_conventions_channels():
-    """Test swap_angular_conventions with channel dimension."""
-    data = torch.randn(2, 3, 5, 2)
-    result = swap_angular_conventions(data)
-    assert result.shape == (2, 5, 3, 2)
-
-
-def test_swap_angular_conventions_identity():
-    """Test that applying swap twice returns original."""
-    data = torch.randn(2, 3, 5)
-    result = swap_angular_conventions(swap_angular_conventions(data))
-
-    assert torch.allclose(result, data)
-    assert result.shape == data.shape
-
-
-# Voxels Tests
-
-def test_bev_from_polar2_basic():
-    """Test basic bev_from_polar2 functionality."""
-    data = _small_polar_data()  # (2,6,8,1)
-    result = bev_from_polar2(data, size=16)
-    assert result.shape == (2, 16, 32, 1)
-    assert result.dtype == data.dtype
-
-
-def test_bev_from_polar2_different_theta():
-    """Test bev_from_polar2 with custom theta range."""
-    data = _small_polar_data()  # (2,6,8,1)
-    result = bev_from_polar2(
-        data, size=8, theta_min=-np.pi / 2, theta_max=np.pi / 2)
-
-    assert result.shape == (2, 8, 16, 1)
-
-
-def test_bev_from_polar2_bool_data():
-    """Test bev_from_polar2 preserves bool dtype."""
-    data = torch.randint(0, 2, (2, 6, 8, 1), dtype=torch.bool)
-    result = bev_from_polar2(data, size=8)
-
-    assert result.dtype == torch.bool
-    assert result.shape == (2, 8, 16, 1)
-
-
-def test_bev_height_from_polar_occupancy_scaled():
-    """Test bev_height_from_polar_occupancy with scaling."""
-    data = _small_polar_occupancy()  # (2,3,6,8)
-    result = bev_height_from_polar_occupancy(data, size=16, scale=True)
-    # If it succeeds, check basic properties
-    assert result.shape == (2, 16, 32)
-    assert result.dtype == torch.float32
-
-
-def test_bev_height_from_polar_occupancy_unscaled_small():
-    """Test bev_height_from_polar_occupancy without scaling, small elevation."""
-    data = _small_polar_occupancy()  # (2,3,6,8) - 3 elevation bins <= 256
-    result = bev_height_from_polar_occupancy(data, size=16, scale=False)
-    assert result.shape == (2, 16, 32)
-    assert result.dtype == torch.uint8
-
-
-def test_bev_height_from_polar_occupancy_unscaled_large():
-    """Test bev_height_from_polar_occupancy with >256 elevation bins."""
-    # Create data with 300 elevation bins
-    data = torch.randint(0, 2, (1, 300, 6, 8), dtype=torch.bool)
-    result = bev_height_from_polar_occupancy(data, size=16, scale=False)
-    assert result.shape == (1, 16, 32)
-    assert result.dtype in [torch.int16, torch.uint8]
-
-
-def test_bev_height_from_polar_occupancy_empty():
-    """Test bev_height_from_polar_occupancy with empty occupancy."""
-    data = torch.zeros(2, 3, 6, 8, dtype=torch.bool)  # All false
-    result = bev_height_from_polar_occupancy(data, size=8, scale=False)
-
-    assert result.shape == (2, 8, 16)
-    assert result.dtype == torch.uint8
-    assert torch.all(result == 0)
-
-
-def test_depth_from_polar_occupancy_basic():
-    """Test basic depth_from_polar_occupancy functionality."""
-    data = _small_polar_occupancy()  # (2,3,6,8) - 8 range bins <= 256
-    result = depth_from_polar_occupancy(data)
-    assert result.shape == (2, 3, 6)
-    assert result.dtype == torch.uint8
-
-
-def test_depth_from_polar_occupancy_with_resize():
-    """Test depth_from_polar_occupancy with resizing."""
-    data = _small_polar_occupancy()  # (2,3,6,8)
-    result = depth_from_polar_occupancy(data, size=(4, 7))  # Non-square target
-    assert result.shape == (2, 4, 7)
-    assert result.dtype == torch.uint8
-
-
-def test_depth_from_polar_occupancy_large_range():
-    """Test depth_from_polar_occupancy with >256 range bins."""
-    data = torch.randint(0, 2, (1, 3, 6, 300), dtype=torch.bool)
-    result = depth_from_polar_occupancy(data)
-    assert result.shape == (1, 3, 6)
-    assert result.dtype == torch.int16
-
-
-def test_depth_from_polar_occupancy_all_empty():
-    """Test depth_from_polar_occupancy with all-empty occupancy."""
-    data = torch.zeros(2, 3, 6, 8, dtype=torch.bool)
-    result = depth_from_polar_occupancy(data)
-
-    assert result.shape == (2, 3, 6)
-    assert torch.all(result == 0)
-
-
-def test_depth_from_polar_occupancy_all_occupied():
-    """Test depth_from_polar_occupancy with all-occupied occupancy."""
-    data = torch.ones(2, 3, 6, 8, dtype=torch.bool)
-    result = depth_from_polar_occupancy(data)
-
-    assert result.shape == (2, 3, 6)
-    assert torch.all(result == 0)
-
-
-# Edge Cases
-
-def test_tile_images_single_image():
-    """Test tile_images with single image."""
-    image = torch.randn(1, 3, 5)
-    result = tile_images(image, cols=1, normalize=True)
-
-    assert result.shape == (3, 5, 3)
-
-
-def test_tile_images_more_cols_than_images():
-    """Test tile_images when cols > batch_size."""
-    images = torch.randn(2, 3, 5)
-    result = tile_images(images, cols=5, normalize=True)  # ncols > images
-    assert result.shape == (3, 10, 3)  # 1 row × 2 actual images
-
-
-def test_swap_angular_conventions_single_pixel():
-    """Test swap_angular_conventions with minimal data."""
-    data = torch.tensor([[[1.0, 2.0], [3.0, 4.0]]])  # (1,2,2)
-    result = swap_angular_conventions(data)
-
-    assert result.shape == (1, 2, 2)
-    # The actual transformation: flip(dims=[1,2]) then transpose(1,2)
-    # [[1,2], [3,4]] -> flip -> [[4,3], [2,1]] -> transpose -> [[4,2], [3,1]]
-    expected = torch.tensor([[[4.0, 2.0], [3.0, 1.0]]])
-    assert torch.allclose(result, expected)
-
-
-def test_bev_from_polar2_minimal_size():
-    """Test bev_from_polar2 with minimal size."""
-    data = torch.randn(1, 4, 4, 1)
-    result = bev_from_polar2(data, size=4)
-
-    assert result.shape == (1, 4, 8, 1)
